+++ build.gradle
@@ -5,8 +5,8 @@ plugins {
 }
 
 group 'net.notfab.lib'
-version '1.2.2'
-description 'Persistence Engine'
+version '1.3'
+description 'Hibernate Wrapper Engine'
 
 sourceCompatibility = targetCompatibility = 1.8
