+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'PersistEngine'
+rootProject.name = 'Hibernitto'
