+++ src/main/java/net/notfab/hibernitto/jpa/JPAEngine.java
@@ -1,23 +1,24 @@
-package net.notfab.lib;
+package net.notfab.hibernitto.jpa;
 
-import net.notfab.lib.entities.SQLFilter;
-import net.notfab.lib.entities.SQLWhere;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
+import net.notfab.hibernitto.core.Dialect;
+import net.notfab.hibernitto.core.entities.SQLFilter;
+import net.notfab.hibernitto.core.entities.SQLWhere;
+import org.hibernate.internal.SessionImpl;
 
 import javax.persistence.Entity;
-import javax.persistence.EntityTransaction;
+import javax.persistence.EntityManager;
 import javax.persistence.Query;
 import javax.persistence.Table;
 import java.util.List;
 
-public class HibernateEngine implements AutoCloseable {
+@SuppressWarnings({"WeakerAccess", "unchecked", "SqlDialectInspection", "SqlNoDataSourceInspection"})
+public class JPAEngine implements AutoCloseable {
 
-    private Session session;
+    private EntityManager em;
     private Dialect dialect;
 
-    public HibernateEngine(Session session, Dialect dialect) {
-        this.session = session;
+    public JPAEngine(EntityManager em, Dialect dialect) {
+        this.em = em;
         this.dialect = dialect;
     }
 
@@ -26,15 +27,14 @@ public class HibernateEngine implements AutoCloseable {
      *
      * @param object The object.
      */
-    public void persist(Object object) {
-        EntityTransaction transaction = session.getTransaction();
-        transaction.begin();
-        session.saveOrUpdate(object);
-        transaction.commit();
+    public void save(Object object) {
+        em.getTransaction().begin();
+        ((SessionImpl) em).saveOrUpdate(object);
+        em.getTransaction().commit();
     }
 
     /**
-     * Fetches a persisted entity from the database.
+     * Fetches an entity from the database.
      *
      * @param clazz Class of the entity
      * @param o     Filter object(s) or primary key
@@ -45,12 +45,14 @@ public class HibernateEngine implements AutoCloseable {
         if (o instanceof SQLWhere) return this.get(clazz, (SQLWhere) o); // SQLWhere
         if (isArray(o)) return this.get(clazz, (SQLFilter[]) o); // Array of SQLFilters
         if (o instanceof SQLFilter) return this.get(clazz, new SQLFilter[]{(SQLFilter) o}); // SQLFilter
-        T resp = session.find(clazz, o);
+        em.getTransaction().begin();
+        T resp = em.find(clazz, o);
+        em.getTransaction().commit();
         return resp;
     }
 
     /**
-     * Fetches a persisted entity from the database.
+     * Fetches an entity from the database.
      *
      * @param clazz Class of the entity
      * @param where SQLWhere object (filter).
@@ -59,7 +61,7 @@ public class HibernateEngine implements AutoCloseable {
      */
     public <T> T get(Class<T> clazz, SQLWhere where) {
         String queryStr = "SELECT * from " + this.getClassName(clazz) + where.toString();
-        Query query = session.createNativeQuery(queryStr, clazz);
+        Query query = em.createNativeQuery(queryStr, clazz);
         where.prepare(query);
         query.setFirstResult(0);
         query.setMaxResults(1);
@@ -67,17 +69,17 @@ public class HibernateEngine implements AutoCloseable {
     }
 
     /**
-     * Fetches a persisted entity from the database with filters.
+     * Fetches an entity from the database with filters.
      *
      * @param clazz   Class of the entity
-     * @param filters SQL Filters see {@link net.notfab.lib.entities}
+     * @param filters SQL Filters see {@link net.notfab.hibernitto.core.entities}
      * @param <T>     Type
-     * @return Persisted entity if found, null otherwise.
+     * @return Entity if found, null otherwise.
      */
     public <T> T get(Class<T> clazz, SQLFilter... filters) {
         SQLWhere where = new SQLWhere(filters);
         String queryStr = "SELECT * from " + this.getClassName(clazz) + where.toString();
-        Query query = session.createNativeQuery(queryStr, clazz);
+        Query query = em.createNativeQuery(queryStr, clazz);
         where.prepare(query);
         query.setFirstResult(0);
         query.setMaxResults(1);
@@ -87,18 +89,18 @@ public class HibernateEngine implements AutoCloseable {
     // --------------------------------------------------
 
     /**
-     * Fetches a list of persisted entities from the database with filters.
+     * Fetches a list of entities from the database with filters.
      *
      * @param clazz       Class of the entity
      * @param firstResult First result
      * @param maxResults  Max results to return
      * @param where       SQLWhere object (filter)
      * @param <T>         Type
-     * @return Persisted list of entities if found, null otherwise.
+     * @return List of entities if found, null otherwise.
      */
     public <T> List<T> getList(Class<T> clazz, int firstResult, int maxResults, SQLWhere where) {
         String queryStr = "SELECT * from " + this.getClassName(clazz) + where.toString();
-        Query query = session.createNativeQuery(queryStr, clazz);
+        Query query = em.createNativeQuery(queryStr, clazz);
         query = where.prepare(query);
         query.setFirstResult(firstResult);
         query.setMaxResults(maxResults);
@@ -106,55 +108,55 @@ public class HibernateEngine implements AutoCloseable {
     }
 
     /**
-     * Fetches a list of persisted entities from the database with filters.
+     * Fetches a list of entities from the database with filters.
      * <p>
      * Note: This has a default limit of 50 results.
      *
      * @param clazz Class of the entity
      * @param where SQLWhere object (filter)
      * @param <T>   Type
-     * @return Persisted list of entities if found, null otherwise.
+     * @return List of entities if found, null otherwise.
      */
     public <T> List<T> getList(Class<T> clazz, SQLWhere where) {
         return getList(clazz, 0, 50, where);
     }
 
     /**
-     * Fetches a list of persisted entities from the database with filters.
+     * Fetches a list of entities from the database with filters.
      * <p>
      * Note: This has a default limit of 50 results.
      *
      * @param clazz   Class of the entity
-     * @param filters SQL Filters see {@link net.notfab.lib.entities}
+     * @param filters SQL Filters see {@link net.notfab.hibernitto.core.entities}
      * @param <T>     Type
-     * @return Persisted list of entities if found, null otherwise.
+     * @return List of entities if found, null otherwise.
      */
     public <T> List<T> getList(Class<T> clazz, SQLFilter... filters) {
         return getList(clazz, 0, 50, filters);
     }
 
     /**
-     * Fetches a list of persisted entities from the database with filters.
+     * Fetches a list of entities from the database with filters.
      *
      * @param clazz      Class of the entity
      * @param maxResults Max results to return
-     * @param filters    SQL Filters see {@link net.notfab.lib.entities}
+     * @param filters    SQL Filters see {@link net.notfab.hibernitto.core.entities}
      * @param <T>        Type
-     * @return Persisted list of entities if found, null otherwise.
+     * @return List of entities if found, null otherwise.
      */
     public <T> List<T> getList(Class<T> clazz, int maxResults, SQLFilter... filters) {
         return getList(clazz, 0, maxResults, filters);
     }
 
     /**
-     * Fetches a list of persisted entities from the database with filters.
+     * Fetches a list of entities from the database with filters.
      *
      * @param clazz       Class of the entity
      * @param firstResult First result
      * @param maxResults  Max results to return
-     * @param filters     SQL Filters see {@link net.notfab.lib.entities}
+     * @param filters     SQL Filters see {@link net.notfab.hibernitto.core.entities}
      * @param <T>         Type
-     * @return Persisted list of entities if found, null otherwise.
+     * @return List of entities if found, null otherwise.
      */
     public <T> List<T> getList(Class<T> clazz, int firstResult, int maxResults, SQLFilter... filters) {
         return getList(clazz, firstResult, maxResults, new SQLWhere(filters));
@@ -167,7 +169,7 @@ public class HibernateEngine implements AutoCloseable {
      * @return Query.
      */
     public Query createNativeQuery(String query) {
-        return session.createNativeQuery(query);
+        return em.createNativeQuery(query);
     }
 
     /**
@@ -178,7 +180,7 @@ public class HibernateEngine implements AutoCloseable {
      * @return Query.
      */
     public Query createNativeQuery(String query, Class<?> clazz) {
-        return session.createNativeQuery(query, clazz);
+        return em.createNativeQuery(query, clazz);
     }
 
     /**
@@ -186,12 +188,11 @@ public class HibernateEngine implements AutoCloseable {
      *
      * @param object The entity.
      */
-    public void delete(Object object) {
+    public void remove(Object object) {
         if (object == null) return;
-        Transaction transaction = session.getTransaction();
-        transaction.begin();
-        session.remove(session.merge(object));
-        transaction.commit();
+        em.getTransaction().begin();
+        em.remove(em.merge(object));
+        em.getTransaction().commit();
     }
 
     /**
@@ -200,7 +201,7 @@ public class HibernateEngine implements AutoCloseable {
      * @param o The entity.
      */
     public void refresh(Object o) {
-        session.refresh(o);
+        em.refresh(o);
     }
 
     /**
@@ -209,7 +210,7 @@ public class HibernateEngine implements AutoCloseable {
      * @param o The entity.
      */
     public void detach(Object o) {
-        session.detach(o);
+        em.detach(o);
     }
 
     /**
@@ -219,14 +220,14 @@ public class HibernateEngine implements AutoCloseable {
      * @return unwrapped entity.
      */
     public <T> T unwrap(Class<T> o) {
-        return session.unwrap(o);
+        return em.unwrap(o);
     }
 
     /**
-     * Closes the session.
+     * Closes the connection pool and entity manager.
      */
     public void close() {
-        this.session.close();
+        this.em.close();
     }
 
     private boolean isArray(Object obj) {
@@ -241,6 +242,10 @@ public class HibernateEngine implements AutoCloseable {
         }
         Entity entity = clazz.getDeclaredAnnotation(Entity.class);
         return entity.name().equals("") ? clazz.getSimpleName() : entity.name();
+    }
+
+    public Dialect getDialect() {
+        return dialect;
     }
 
 }
