+++ src/main/java/net/notfab/hibernitto/jpa/JPAFactory.java
@@ -1,23 +1,25 @@
-package net.notfab.lib;
+package net.notfab.hibernitto.jpa;
+
+import net.notfab.hibernitto.core.Dialect;
 
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
 import java.io.Closeable;
 import java.util.Properties;
 
-public class PersistenceFactory implements Closeable {
+public class JPAFactory implements Closeable {
 
     private final String name;
     private final EntityManagerFactory entityManagerFactory;
     private final Dialect dialect;
 
-    public PersistenceFactory(String name, Dialect dialect) {
+    public JPAFactory(String name, Dialect dialect) {
         this.name = name;
         this.dialect = dialect;
         this.entityManagerFactory = Persistence.createEntityManagerFactory(name);
     }
 
-    public PersistenceFactory(String name, Dialect dialect, Properties properties) {
+    public JPAFactory(String name, Dialect dialect, Properties properties) {
         this.name = name;
         this.dialect = dialect;
         this.entityManagerFactory = Persistence.createEntityManagerFactory(name, properties);
@@ -27,8 +29,8 @@ public class PersistenceFactory implements Closeable {
         return name;
     }
 
-    public PersistEngine getEngine() {
-        return new PersistEngine(this.entityManagerFactory.createEntityManager(), this.dialect);
+    public JPAEngine getEngine() {
+        return new JPAEngine(this.entityManagerFactory.createEntityManager(), this.dialect);
     }
 
     @Override
@@ -36,8 +38,8 @@ public class PersistenceFactory implements Closeable {
         this.entityManagerFactory.close();
     }
 
-    public static PersistenceFactoryBuilder builder(Dialect dialect) {
-        return new PersistenceFactoryBuilder(dialect);
+    public static JPAFactoryBuilder builder(Dialect dialect) {
+        return new JPAFactoryBuilder(dialect);
     }
 
 }
